requirements.txt
streamlit
openai>=1.55.0
requests==2.31.0
python-dotenv==1.0.0
httpx>=0.27.0
typing-extensions>=4.9.0